requirements.txt
crewai==0.28.8
crewai-tools==0.1.7
python-dotenv==1.0.1
fpdf2==2.7.8